refactor(cli): share telemetry helper logic
